# server/docs/swagger.yaml
basePath: /api/v1
definitions:
  model.Room:
    properties:
      Description:
        example: Sample description
        type: string
      IsPrivate:
        example: false
        type: boolean
      Name:
        example: SampleName
        type: string
    type: object
info:
  contact: {}
  title: StreamSync API
  version: "1.0"
paths:
  /rooms/:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Room'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Retrieves all rooms
  /rooms/{id}:
    get:
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Room'
      security:
      - ApiKeyAuth: []
      summary: Retrieves room based on given ID
  /rooms/create:
    post:
      parameters:
      - description: Add Room
        in: body
        name: room
        required: true
        schema:
          $ref: '#/definitions/model.Room'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Room'
      security:
      - ApiKeyAuth: []
      summary: Creates room
  /rooms/delete{id}:
    delete:
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Room'
      security:
      - ApiKeyAuth: []
      summary: Deletes room
  /rooms/update{id}:
    put:
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Room'
      security:
      - ApiKeyAuth: []
      summary: Updates room
swagger: "2.0"